FEAT: Allow weighting for penalized_poly
The polynomial.penalized_poly function now allows weights to be used. Added tests comparing the weighted results to that of the poly and modpoly weighted results, which should be equal the penalized_poly results when using specific values, to ensure that the weighting for penalized_poly is correct. Also made the default threshold value scale with the data better.