FEAT: Significantly improved loess speed and memory
Significantly reduced both the calculation time and the memory usage of polynomial.loess. Added a conserve_memory parameter that will recalculate the kernels each time, which is slower but uses very little memory. Allow inputting weights into loess, added use_original  kwarg for thresholding to be the same as the modpoly and imodpoly functions, and added return_coef kwarg to allow returning the polynomial coefficients for each x-value to recreate the fitting polynomial, to match all other polynomial functions.