FEAT: Slightly speed up loess kernel calc
Changed the logic for shifting the left and right indices of the window, following section 5.1 in Cleveland's 1979 paper. This allows only calculating the difference for the actual windows values, which is slightly faster. Store windows as a 2d numpy array rather than a list of slices, which will be compatible when switching to numba.