FEAT: Allow using numba for loess calculation to improve speed
Used numba's jit decorator to improve speed of loess calculation and allow doing the calculations in parallel. Use numpy's solve rather than lstsq for loess since solve is faster and has similar accuracy. Set default `conserve_memory` to True, since it is just as fast as False when numba is installed and is safer. Increased the number of digits for the _median_absolute_values divisor to not artificially reduce accuracy. Added a test comparing the output of statsmodel's lowess to ensure the loess function is accurate.