OTHER: change behavior of loess interpolation
Renamed the variable skip_dx to delta to be consistent with both Cleveland's original Fortran lowess program and statsmodels's implementation. Also changed the behavior of delta to fit the last value within x_last_fit + delta, for the same reasons. Added a test to compare with statsmodels when using delta. Also now sort the weight array based on the x-sorting.