FEAT: Improve spline.mixture_model numerical stability
The y-values are scaled between -1 and 1 so that the residual is also ~ -1 to 1, making the histogram better when fitting values where y covers a large scale and improving the fit stability. Use log10(sigma) in the fit so that sigma is not bound and can cover larger scales more easily.