DOCS: Fix mpls's plot in docs
The plot of mpls for the data with positive and negative peaks was producing an error since there were few points with a weight not equal to 0, so set p to a non-zero value so that all points have weighting. Did the same for mpspline, which could also experience a similar issue.