OTHER: Make convenience functions for whittaker smoothing
The function utils._pentapy_solver now tells the current utils.PENTAPY_SOLVER value, which allows not reading the value directly from utils. Made a dedicated function for whittaker smoothing in _algorithm_setup, so that it can be called for one-off computations, like mpls and the soon to be added fabc.