FIX: Fixed edge cases for utils.padded_convolve
Fixed the output of padded_convolve when the kernel was even shaped (which never happens in actual application in pybaselines) or larger than the data. Switched to scipy's convolve, which uses fft for larger inputs, which is significantly faster. ~5-10% slower for small inputs (<10000), but ~50-90% faster for larger inputs compared to numpy's convolve. Added various tests to ensure the convolution output is the same shape as the input.