DEPR: Deprecated array half_windows for rolling_ball
The array moving min/max functions were valid, but when combined in the opening/closing functions, would produce invalid results. For example, an array opening would produce values greater than the input data, which should not be allowed by the actual morphological opening. The production of invalid data is not expected by the user and is not suitable for fitting baselines, and allowing only integers greatly simplifies the morphological functions. Thus, will remove array half-window support for rolling_ball in version 0.8.0.