FIX: Can now pass weights to optimize_extended_range
If weights are input into optimize_extended_range as keyword arguments, the weights are now correctly sorted and padded to account for the added portions before using in the fitting function. Also slightly sped up the function by only minimizing the sum of squares rather than the root mean sum of squares since they are minimized the same way.