BREAK: Count first and last knots as inner knots for spline basis
The first and last knots are now included as inner knots when creating the spline basis rather than counting them as the first outer knots. Most other packages do it this way (scikit-learn and scipy's new BSpline.design_matrix), so it is more convenient to follow their approach. Not a huge change to the baseline algorithms. To retain previous behavior, use  previous `num_knots` + 2.