OTHER: Switch difference penalty diagonals from upper to lower
Switch from using the upper bands to the lower bands for the finite different penalty matrix. This is in anticipation of using lower bands for the new spline methods. Lower should also be slightly faster than upper for the cholesky decomposition, but the impact is probably minor. Added new functions for computing the penalty bands of any order and for the penalty bands of the third difference order.