FEAT: Switched smooth to class api
Added all smooth functions as methods for the new Smooth class. Added `x_data` kwarg to all functions that previously did not take it. Switched to using np.gradient instead of np.diff for swima for less noisy calc. Replaced all smooth tests with the new test class. Will make the functions as aliases to the class methods in a later commit.