FEAT: Switched optimizers to class api
Added all optimizers functions as methods for the new Optimizers class. Fixed inputting `alpha` for optimize_extended_range with 'aspls' method. Added `x_data` kwarg to all functions that previously did not take it. Replaced all optimizers tests with the new test class. Will make the functions as aliases to the class methods in a later commit. Also need to rethink how to do the sorting of the input data and output baselines for optimizers.