MAINT: Change `assume_sorted` behavior
Input x and z values are checked to be in ascending order even if assume_sorted is True so that individual methods don't have to check for ordering. Also added option to _Algorithm[2D].register to signify if a method needs unique x or z values and allow only having to do that check once per instance.